requirements.txt
requests
pandas
openpyxl
python-docx
PyPDF2
